% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{add_flank_to_variants}
\alias{add_flank_to_variants}
\title{Uses a genome object to find context and add it to the variant table}
\usage{
add_flank_to_variants(bay, g, flank_start, flank_end, build_table = TRUE)
}
\arguments{
\item{bay}{Input samples}

\item{g}{Genome object used for finding variant context}

\item{flank_start}{Start of flank area to add, can be positive or negative}

\item{flank_end}{End of flank area to add, can be positive or negative}

\item{build_table}{Automatically build a table using the annotation and add
it to the bagel}
}
\description{
Uses a genome object to find context and add it to the variant table
}
\examples{
bay <- readRDS(system.file("testdata", "bagel_snv96_tiny.rds", package = "BAGEL"))
g <- select_genome("38")
add_flank_to_variants(bay, g, 1, 2)
add_flank_to_variants(bay, g, -2, -1)
}
