% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{add_variant_type}
\alias{add_variant_type}
\title{Generates a variant type table}
\usage{
add_variant_type(tab)
}
\arguments{
\item{tab}{Input variant table}
}
\description{
Generates a variant type table
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
BAGEL:::add_variant_type(bay@variants)
}
\keyword{internal}
