% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_breakpoints}
\alias{all_breakpoints}
\title{Pan-Cancer Breakpoint Data from BISCUT Analysis}
\format{
A data frame with variables:
\describe{
\item{peak_id}{Unique identifier for each breakpoint peak}
\item{arm}{Chromosome arm (e.g., "1p", "1q")}
\item{direction}{Peak direction: "amp" (amplification) or "del" (deletion)}
\item{Chr}{Chromosome number}
\item{Peak.Start}{Peak start position (bp)}
\item{Peak.End}{Peak end position (bp)}
\item{n_events}{Number of events contributing to the peak}
\item{ks_stat}{Kolmogorov-Smirnov statistic}
\item{ks_p}{K-S test p-value}
\item{log10_ks_p}{-log10(p-value)}
\item{cancer_type}{Cancer type from TCGA}
\item{...}{Additional BISCUT analysis columns}
}
}
\source{
TCGA pan-cancer segmentation data processed with BISCUT algorithm
}
\usage{
all_breakpoints
}
\description{
Complete BISCUT breakpoint detection results across 29 cancer types from TCGA.
This dataset contains detailed peak information for chromosome arm breakpoints
identified using the BISCUT algorithm.
}
\keyword{datasets}
