% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aneu}
\alias{aneu}
\title{aneu}
\format{
\subsection{aneu}{

\describe{
\item{direction}{1, -1, -2, 2}
}
}
}
\source{
NA
}
\usage{
aneu
}
\description{
Aneuploidy dataframe
}
\keyword{datasets}
