% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_replication_strand}
\alias{annotate_replication_strand}
\title{Add replication strand annotation to SBS variants based on bedgraph file}
\usage{
annotate_replication_strand(bay, rep_range, build_table = TRUE)
}
\arguments{
\item{bay}{Input bagel}

\item{rep_range}{A GRanges object with replication timing as metadata}

\item{build_table}{Automatically build a table from this annotation}
}
\value{
None
}
\description{
Add replication strand annotation to SBS variants based on bedgraph file
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
annotate_replication_strand(bay, BAGEL::rep_range)
}
