% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number-analysis.R
\name{annotate_segments_with_arms}
\alias{annotate_segments_with_arms}
\title{Annotate Segments with Arms Using Genomic Overlaps}
\usage{
annotate_segments_with_arms(segments, arm_definitions)
}
\arguments{
\item{segments}{Segmentation data}

\item{arm_definitions}{Arm definitions}
}
\value{
Segments annotated with arm assignments
}
\description{
Annotate Segments with Arms Using Genomic Overlaps
}
\keyword{internal}
