% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{annotate_snv_strand}
\alias{annotate_snv_strand}
\title{Add strand annotation to SNV variants}
\usage{
annotate_snv_strand(bay)
}
\arguments{
\item{bay}{Input bagel}
}
\description{
Add strand annotation to SNV variants
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
annotate_snv_strand(bay)
}
