% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_transcript_strand}
\alias{annotate_transcript_strand}
\title{Add transcript strand annotation to SBS variants (defined in genes only)}
\usage{
annotate_transcript_strand(bay, genome_build, build_table = TRUE)
}
\arguments{
\item{bay}{Input bagel}

\item{genome_build}{Which genome build to use: hg19, hg38, or a custom TxDb
object}

\item{build_table}{Automatically build a table from this annotation}
}
\value{
None
}
\description{
Add transcript strand annotation to SBS variants (defined in genes only)
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
annotate_transcript_strand(bay, 19)
}
