% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{annotate_variant_type}
\alias{annotate_variant_type}
\title{Annotate variants with variant type ("SNV", "INS", "DEl", "DBS")}
\usage{
annotate_variant_type(bay)
}
\arguments{
\item{bay}{Input bagel}
}
\description{
Annotate variants with variant type ("SNV", "INS", "DEl", "DBS")
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
annotate_variant_type(bay)
}
