% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateArms.R
\name{annotateArms}
\alias{annotateArms}
\title{Annotate Segments with Chromosome Arms}
\usage{
annotateArms(segments, genome = "hg38", cytoband = NULL)
}
\arguments{
\item{segments}{A data frame containing segment information with columns:
\itemize{
\item Sample: Identifier for the sample
\item Chromosome: Chromosome number
\item Start: Start position of the segment
\item End: End position of the segment
\item Num_Probes: Number of probes in the segment
\item Segment_Mean: Mean value for the segment
\item Arm: An initialized column that will be filled with the corresponding arm
}}

\item{arms}{A data frame containing chromosome arm information with columns:
\itemize{
\item Chromosome: Chromosome number
\item Arm: Arm identifier (e.g., "1p", "1q")
\item Start: Start position of the chromosome arm
\item End: End position of the chromosome arm
}}
}
\value{
A data frame with the original segments, now including the
corresponding Arm for each segment. Segments that cross the centromere
are split into separate p-arm and q-arm segments, with probes distributed
according to their length in each arm.
}
\description{
This function annotates segments with their corresponding chromosome arms
based on the provided arm information. It checks for overlaps between
segment start and end positions and the defined arms for each chromosome.
If a segment crosses the centromere, it is split into two daughter segments:
one for the p-arm and one for the q-arm. These daughter segments retain the
same log2 ratios, and the number of probes is split based on the proportion
of the original segment that lies in each arm.
}
