% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_prediction.R
\name{auto_subset_sigs}
\alias{auto_subset_sigs}
\title{Automatic filtering of inactive signatures}
\usage{
auto_subset_sigs(
  bagel,
  table_name,
  signature_res,
  min_exists = 0.05,
  proportion_samples = 0.25,
  rare_exposure = 0.4
)
}
\arguments{
\item{bagel}{Input samples to predit signature weights}

\item{table_name}{Name of table used for posterior prediction (e.g. SNV96)}

\item{signature_res}{Signatures to automatically subset from for prediction}

\item{min_exists}{Threshold to consider a signature active in a sample}

\item{proportion_samples}{Threshold of samples to consider a signature
active in the cohort}

\item{rare_exposure}{A sample will be considered active in the cohort if at
least one sample has more than this threshold proportion}
}
\value{
Results a result object containing automatically subset signatures
and corresponding sample weights
}
\description{
Automatic filtering of inactive signatures
}
\examples{
bay <- readRDS(system.file("testdata", "bagel_annot.rds", package = "BAGEL"))
auto_subset_sigs(bay, "SNV96", BAGEL::cosmic_v2_sigs)
}
