% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{auto_to_bagel_dt}
\alias{auto_to_bagel_dt}
\title{Chooses the correct function to load in input based on class or file
extension}
\usage{
auto_to_bagel_dt(
  input,
  name = NULL,
  filter = TRUE,
  only_snp = TRUE,
  extra_fields = NULL,
  auto_fix_errors = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{input}{VCF, MAF, file.vcf, file.maf, or list of inputs}

\item{name}{Optional name for better error-reporting}

\item{filter}{Filter to only passed variants}

\item{only_snp}{Filter only non-snp variants}

\item{extra_fields}{Which additional fields to extract}

\item{auto_fix_errors}{Attempt to automatically fix file formatting errors}

\item{verbose}{Show file list progress}
}
\value{
Returns a data.table of variants from a vcf
}
\description{
Chooses the correct function to load in input based on class or file
extension
}
\examples{
luad_vcf_file <- system.file("testdata", "public_LUAD_TCGA-97-7938.vcf",
  package = "BAGEL")
luad <- BAGEL::auto_to_bagel_dt(input = luad_vcf_file)

luad_vcf_file <- system.file("testdata", "public_LUAD_TCGA-97-7938.vcf",
  package = "BAGEL")
luad_vcf <- VariantAnnotation::readVcf(luad_vcf_file)
luad_vcf_name <- basename(luad_vcf_file)
luad <- BAGEL::auto_to_bagel_dt(input = luad_vcf, name = luad_vcf_name)

maf_file=system.file("testdata", "public_TCGA.LUSC.maf", package = "BAGEL")
maf = maftools::read.maf(maf_file)
dt = BAGEL::maf_to_dt(maf)
maf_dt = BAGEL::auto_to_bagel_dt(input = dt, filter = FALSE)

melanoma_vcfs <- list.files(system.file("testdata", package = "BAGEL"),
  pattern = glob2rx("*SKCM*vcf"), full.names = TRUE)
melanoma <- BAGEL::auto_to_bagel_dt(input = melanoma_vcfs)
}
