% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\docType{class}
\name{bagel-class}
\alias{bagel-class}
\title{The primary object for BayeSig that contains all samples and results}
\description{
The primary object for BayeSig that contains all samples and results
}
\section{Slots}{

\describe{
\item{\code{samples}}{A list of sample objects containing sample-level information}

\item{\code{prop_table}}{Summary tables with motif counts normalized by sample sums}

\item{\code{counts_table}}{Summary tables with unnormalized motif counts}
}}

