% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel-workflow.R
\name{bagel-workflow}
\alias{bagel-workflow}
\title{BAGEL v2.0 Workflow Functions}
\description{
Complete workflow functions for BAGELv2 analysis including data loading,
matrix creation, and report generation.
}
