% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bagel_log}
\alias{bagel_log}
\title{Internal Logging Function}
\usage{
bagel_log(message, level = "INFO")
}
\arguments{
\item{message}{Character, message to log}

\item{level}{Character, log level}
}
\description{
Internal Logging Function
}
\keyword{internal}
