% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bladder}
\alias{bladder}
\title{bladder}
\format{
\subsection{bladder}{

\describe{
\item{chr}{chromosome}
\item{arm}{chromosome arm}
\item{cytoband}{cytobands}
\item{start}{start base of breakpoint region}
\item{end}{end base of breakpoint region}
\item{direction}{direction of the breakpoint conferring advantage}
\item{telcent}{direction of the breakpoint calculated, telomere to centromere or opposite}
\item{id}{combined id of each breakpoint event}
}
}
}
\source{
\url{https://www.github.com/PollyHung/BAGEL/docs/bladder.txt}
}
\usage{
bladder
}
\description{
Breakpoint File for TCGA Bladder Cancer
}
\keyword{datasets}
