% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breakpoint_ids}
\alias{breakpoint_ids}
\title{breakpoint_ids}
\format{
\subsection{breakpoint_ids}{

\describe{
\item{breakpoint_ids}{breakpoint_ids}
}
}
}
\source{
\url{https://www.github.com/PollyHung/BAGEL/docs/breakpoint_ids.txt}
}
\usage{
breakpoint_ids
}
\description{
All Breakpoint IDS
}
\keyword{datasets}
