% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{build_custom_table}
\alias{build_custom_table}
\title{Builds a custom table from specified user variants}
\usage{
build_custom_table(
  bay,
  variant_annotation,
  name,
  description = NA,
  data_factor = NA,
  return_instead = FALSE
)
}
\arguments{
\item{bay}{Input samples}

\item{variant_annotation}{User column to use for building table}

\item{name}{Table name to refer to (must be unique)}

\item{description}{Optional description of the table content}

\item{data_factor}{Full set of table values, in case some are missing from
the data. If NA, a superset of all available unique data values will be used}

\item{return_instead}{Instead of adding to bagel object, return the created
table}
}
\description{
Builds a custom table from specified user variants
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
g <- select_genome("38")
annotate_transcript_strand(bay, "19", build_table = FALSE)
build_custom_table(bay, "Transcript_Strand", "Transcript_Strand",
data_factor = factor(c("T", "U")))
}
