% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint-similarity-analysis.R
\name{calculate_arm_lengths_internal}
\alias{calculate_arm_lengths_internal}
\title{Internal function to calculate arm lengths from chromosome coordinates}
\usage{
calculate_arm_lengths_internal(chr_coords)
}
\arguments{
\item{chr_coords}{Chromosome coordinates from BISCUT}
}
\value{
Data frame with arm length information
}
\description{
Internal function to calculate arm lengths from chromosome coordinates
}
\keyword{internal}
