% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number-analysis.R
\name{calculate_residuals}
\alias{calculate_residuals}
\title{Calculate Residuals from Background Distribution}
\usage{
calculate_residuals(arm_summaries, background_dist)
}
\arguments{
\item{arm_summaries}{Arm-level summaries}

\item{background_dist}{Background distribution parameters}
}
\value{
Residuals data frame
}
\description{
Calculate Residuals from Background Distribution
}
\keyword{internal}
