% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint-similarity-analysis.R
\name{calculate_similarity_statistics}
\alias{calculate_similarity_statistics}
\title{Internal function to calculate similarity statistics}
\usage{
calculate_similarity_statistics(combined_data, dataset_names)
}
\arguments{
\item{combined_data}{Combined breakpoint data}

\item{dataset_names}{Names of the datasets}
}
\value{
List with similarity statistics
}
\description{
Internal function to calculate similarity statistics
}
\keyword{internal}
