% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSegs-fixed.R
\name{check_telcent_alignment}
\alias{check_telcent_alignment}
\title{Check Telomere/Centromere Alignment}
\usage{
check_telcent_alignment(segments, coord, arm_type, telcent)
}
\arguments{
\item{segments}{Classified segments}

\item{coord}{Arm coordinates}

\item{arm_type}{p or q arm}

\item{telcent}{tel or cent}
}
\value{
Logical indicating if alignment passes
}
\description{
Check Telomere/Centromere Alignment
}
\keyword{internal}
