% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint-similarity-analysis.R
\name{combine_and_process_data}
\alias{combine_and_process_data}
\title{Internal function to combine and process breakpoint data}
\usage{
combine_and_process_data(reference_data, query_data, arm_lengths)
}
\arguments{
\item{reference_data}{Reference dataset}

\item{query_data}{Query dataset}

\item{arm_lengths}{Arm length information}
}
\value{
Combined and processed data frame
}
\description{
Internal function to combine and process breakpoint data
}
\keyword{internal}
