% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{compare_cosmic_v2}
\alias{compare_cosmic_v2}
\title{Compare a result object to COSMIC V2 SNV Signatures (combination whole-exome
and whole-genome)}
\usage{
compare_cosmic_v2(
  result,
  threshold = 0.9,
  result_name = deparse(substitute(result))
)
}
\arguments{
\item{result}{Result to compare}

\item{threshold}{threshold for similarity}

\item{result_name}{title for plot user result signatures}
}
\value{
Returns the comparisons
}
\description{
Compare a result object to COSMIC V2 SNV Signatures (combination whole-exome
and whole-genome)
}
\examples{
res <- readRDS(system.file("testdata", "res.rds", package = "BAGEL"))
compare_cosmic_v2(res, threshold = 0.8)
}
