% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{compare_cosmic_v3}
\alias{compare_cosmic_v3}
\title{Compare a result object to COSMIC V2 SNV Signatures (combination whole-exome
and whole-genome)}
\usage{
compare_cosmic_v3(
  result,
  variant_class,
  sample_type,
  threshold = 0.9,
  result_name = deparse(substitute(result))
)
}
\arguments{
\item{result}{Result to compare}

\item{variant_class}{Compare to SNV, DBS, or Indel}

\item{sample_type}{exome (SNV only) or genome}

\item{threshold}{threshold for similarity}

\item{result_name}{title for plot user result signatures}
}
\value{
Returns the comparisons
}
\description{
Compare a result object to COSMIC V2 SNV Signatures (combination whole-exome
and whole-genome)
}
\examples{
res <- readRDS(system.file("testdata", "res.rds", package = "BAGEL"))
compare_cosmic_v3(res, "SNV", "genome", threshold = 0.8)
}
