% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{compare_results}
\alias{compare_results}
\title{Compare two result files or input one to compare to COSMIC}
\usage{
compare_results(result, other_result = cosmic_result, threshold = 0.9,
  what = "bestpairs")
}
\arguments{
\item{result}{Result to compare}

\item{other_result}{Second result, leave blank to use cosmic results}

\item{threshold}{threshold for similarity}

\item{what}{Type of comparison (default is only best pairs)}
}
\value{
Returns the comparisons
}
\description{
Compare two result files or input one to compare to COSMIC
}
\examples{
res <- readRDS(system.file("testdata", "res.rds", package = "BAGEL"))
compare_results(res, threshold = 0.8)
}
