% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_legacy_breakpoints}
\alias{convert_legacy_breakpoints}
\title{Convert Legacy Breakpoint Format to New Format}
\usage{
convert_legacy_breakpoints(legacy_breakpoints)
}
\arguments{
\item{legacy_breakpoints}{List with tel_bound element (legacy format)}
}
\value{
Standardized breakpoint data frame
}
\description{
Convert Legacy Breakpoint Format to New Format
}
\keyword{internal}
