% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\docType{class}
\name{count_table-class}
\alias{count_table-class}
\title{Object containing the count table matrices, their names and descriptions
that we generated by provided and by user functions. These are used to
discover and infer signatures and exposures.}
\description{
Object containing the count table matrices, their names and descriptions
that we generated by provided and by user functions. These are used to
discover and infer signatures and exposures.
}
\section{Slots}{

\describe{
\item{\code{table_list}}{A list of data.tables with counts}

\item{\code{table_name}}{The parameters the result grid was created using}

\item{\code{description}}{A summary table of the result objects in result_list
a list of lists. The nested lists created combined (rbind) tables, and the
tables at the first list level are modelled independantly. Combined tables
must be named.
list("tableA", comboTable = list("tableC", "tableD"))}
}}

