% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\docType{class}
\name{Count_Tables-class}
\alias{Count_Tables-class}
\title{Object containing the result objects generated from the combination of
annotations and a range of k values}
\description{
Object containing the result objects generated from the combination of
annotations and a range of k values
}
\section{Slots}{

\describe{
\item{\code{table_list}}{A list of data.tables with counts}

\item{\code{table_name}}{The parameters the result grid was created using}

\item{\code{description}}{A summary table of the result objects in result_list
a list of lists. The nested lists created combined (rbind) tables, and the
tables at the first list level are modelled independantly. Combined tables
must be named.
list("tableA", comboTable = list("tableC", "tableD"))}
}}

