% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{create_bagel}
\alias{create_bagel}
\title{Creates a bagel object from a variant table}
\usage{
create_bagel(dt, extra_fields = NULL)
}
\arguments{
\item{dt}{data.table input by user or vcf or maf functions to be filtered}

\item{extra_fields}{Which additional fields to extract}
}
\value{
Returns a data.table of variants from a maf
}
\description{
Add description
}
\examples{
maf_file=system.file("testdata", "public_TCGA.LUSC.maf", package = "BAGEL")
maf = maftools::read.maf(maf_file)
dt = maf_to_dt(maf)
maf_dt = create_bagel(dt = dt)
}
