% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint-similarity-analysis.R
\name{create_correlation_plot}
\alias{create_correlation_plot}
\title{Internal function to create correlation plot}
\usage{
create_correlation_plot(matched_data, stat_result, plot_file, dataset_names)
}
\arguments{
\item{matched_data}{Matched breakpoint data}

\item{stat_result}{Statistical results}

\item{plot_file}{Output file path}

\item{dataset_names}{Dataset names}
}
\description{
Internal function to create correlation plot
}
\keyword{internal}
