% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_dbs_table}
\alias{create_dbs_table}
\title{Creates and adds a table for standard doublet base subsitution (DBS)}
\usage{
create_dbs_table(bay, overwrite = overwrite)
}
\arguments{
\item{bay}{Input bagel}

\item{overwrite}{Overwrite existing count table}
}
\description{
Creates and adds a table for standard doublet base subsitution (DBS)
}
