% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint-similarity-analysis.R
\name{create_matched_ideograms}
\alias{create_matched_ideograms}
\title{Internal function to create matched chromosome ideograms}
\usage{
create_matched_ideograms(matched_data, ideogram_file, output_dir)
}
\arguments{
\item{matched_data}{Matched breakpoint data}

\item{ideogram_file}{Output file path}

\item{output_dir}{Output directory}
}
\description{
Internal function to create matched chromosome ideograms
}
\keyword{internal}
