% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_sbs192_table}
\alias{create_sbs192_table}
\title{Uses a genome object to find context and generate standard SBS192 table
using transcript strand}
\usage{
create_sbs192_table(bay, g, strand_type, overwrite = FALSE)
}
\arguments{
\item{bay}{Input samples}

\item{g}{A \linkS4class{BSgenome} object indicating which genome
reference the variants and their coordinates were derived from.}

\item{strand_type}{Transcript_Strand or Replication_Strand}

\item{overwrite}{Overwrite existing count table}
}
\description{
Uses a genome object to find context and generate standard SBS192 table
using transcript strand
}
