% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{create_snv96_table}
\alias{create_snv96_table}
\title{Uses a genome object to find context and generate tables for input samples}
\usage{
create_snv96_table(bay, g)
}
\arguments{
\item{bay}{Input samples}

\item{g}{Genome object used for finding variant context}
}
\description{
Uses a genome object to find context and generate tables for input samples
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
g <- select_genome("38")
create_snv96_table(bay, g)
}
