% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{create_tables}
\alias{create_tables}
\title{Uses a genome object to find context and generate tables for input samples}
\usage{
create_tables(bay, g)
}
\arguments{
\item{bay}{Input samples}

\item{g}{Genome object used for finding variant context}
}
\value{
Returns 96 motif summary tables
}
\description{
Uses a genome object to find context and generate tables for input samples
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "MotifSig"))
g <- select_genome("38")
create_tables(bay, g)
}
