% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCuts.R
\name{createCuts}
\alias{createCuts}
\title{Create Cuts from Genomic Segments}
\usage{
createCuts(
  segments = segs,
  genome = "hg19",
  cutoff = 0.25,
  result_dir = "example/pooledOV/",
  cytoband = NULL
)
}
\arguments{
\item{segments}{A data frame containing genomic segments. Defaults to \code{segs}.}

\item{genome}{A string specifying the genome version (e.g., "hg38", "hg19"). Default is "hg19".}

\item{cutoff}{A numeric value used as a threshold for filtering segments. Default is 0.25.}

\item{result_dir}{A string specifying the directory to save output files. Default is "example/pooledOV/".}

\item{cytoband}{Optional; a path to a custom cytoband CSV file. If provided, overrides the default genome coordinates.}
}
\value{
NULL; the function performs processing and saves output files to the specified result directory.
}
\description{
This function processes genomic segments to create necessary cuts based on specified parameters.
It loads arm-level coordinates for the specified genome and filters segments accordingly.
Additionally, it generates background files for each lineage.
}
\examples{
# Example usage
createCuts(segments = my_segments, genome = "hg38", cutoff = 0.3, result_dir = "output/")

}
