% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoband.hg38}
\alias{cytoband.hg38}
\title{cytoband.hg38}
\format{
\subsection{cytoband.hg38}{

\describe{
\item{chr}{chromosome}
\item{arm}{chromosome arms}
\item{start}{arm start}
\item{end}{arm end}
}
}
}
\source{
\url{https://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/}
}
\usage{
cytoband.hg38
}
\description{
Cytoband version hg38, with centromeres
}
\keyword{datasets}
