% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoband.T2T}
\alias{cytoband.T2T}
\title{cytoband.T2T}
\format{
\subsection{cytoband.T2T}{

\describe{
\item{chr}{chromosome}
\item{arm}{chromosome arms}
\item{start}{arm start}
\item{end}{arm end}
}
}
}
\source{
\url{https://github.com/marbl/CHM13?tab=readme-ov-file}
}
\usage{
cytoband.T2T
}
\description{
Cytoband version T2T, with centromeres
}
\keyword{datasets}
