% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAGEL.R
\name{depression_prob_predict}
\alias{depression_prob_predict}
\title{The prediction function}
\usage{
depression_prob_predict(mcmc, z_new, x_new, seed = 1)
}
\arguments{
\item{mcmc}{the MCMC output given by \link{main_mcmc}}

\item{z_new}{the binary indicator matrix for drug usage information for new visit to predict}

\item{x_new}{the covariates matrix for new visit to predict, it also includes the patient ID (1st column) and visit time (2nd column)}

\item{seed}{the starting number used to generate random numbers}
}
\value{
the predicted probability of reporting depression for each depression item
}
\description{
This function returns the predicted probability of reporting depression for each depression item for a new visit, provided covariates, ART usage information, and previous visit time
}
\seealso{
See \link{main_mcmc} for an example.
}
