% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number-analysis.R
\name{detect_significant_peaks}
\alias{detect_significant_peaks}
\title{Detect Significant Peaks Using Statistical Methods}
\usage{
detect_significant_peaks(residuals, background_dist)
}
\arguments{
\item{residuals}{Residuals from background}

\item{background_dist}{Background distribution}
}
\value{
Peak detection results
}
\description{
Detect Significant Peaks Using Statistical Methods
}
\keyword{internal}
