% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_prediction.R
\name{discover_signatures}
\alias{discover_signatures}
\title{Discovers signatures and weights from a table of counts using NMF}
\usage{
discover_signatures(
  bagel,
  table_name = NULL,
  num_signatures,
  method = "lda",
  seed = 1,
  nstart = 1,
  par_cores = FALSE
)
}
\arguments{
\item{bagel}{A bagel object or counts table}

\item{table_name}{Name of table used for signature discovery}

\item{num_signatures}{Number of signatures to discover, k}

\item{method}{Discovery of new signatures using either LDA or NMF}

\item{seed}{Seed for reproducible signature discovery}

\item{nstart}{Number of independent runs with optimal chosen (lda only)}

\item{par_cores}{Number of parallel cores to use (NMF only)}
}
\value{
Returns a result object with results and input object (if bagel)
}
\description{
Discovers signatures and weights from a table of counts using NMF
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
g <- select_genome("19")
build_standard_table(bay, g, "SBS96", overwrite = TRUE)
discover_signatures(bagel = bay, table_name = "SBS96",
num_signatures = 3, method = "nmf", seed = 12345, nstart = 1)
}
