% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.bagel_env}
\alias{.bagel_env}
\title{Setup BAGEL Environment for Logging}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.bagel_env
}
\description{
Setup BAGEL Environment for Logging
}
\keyword{internal}
