% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{drop_annotation}
\alias{drop_annotation}
\title{Drops a column from the variant table that the user no longer needs}
\usage{
drop_annotation(bay, column_name)
}
\arguments{
\item{bay}{Input samples}

\item{column_name}{Name of column to drop}
}
\description{
Drops a column from the variant table that the user no longer needs
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
drop_annotation(bay, "Variant_Type")
bay
}
