% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{extract_count_tables}
\alias{extract_count_tables}
\title{Extract count tables list from a bagel object}
\usage{
extract_count_tables(bagel)
}
\arguments{
\item{bagel}{bagel to extract count tables list from}
}
\value{
List of count tables objects
}
\description{
Extract count tables list from a bagel object
}
