% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{extract_variants_from_matrix}
\alias{extract_variants_from_matrix}
\title{Extract variants from matrix or data.frame like objects}
\usage{
extract_variants_from_matrix(
  mat,
  chromosome_col = "Chromosome",
  start_col = "Start_Position",
  end_col = "End_Position",
  ref_col = "Tumor_Seq_Allele1",
  alt_col = "Tumor_Seq_Allele2",
  sample_col = "Tumor_Sample_Barcode",
  extra_fields = NULL
)
}
\arguments{
\item{mat}{An object that inherits from classes "matrix" or "data.frame"
Examples include a matrix, data.frame, or data.table.}

\item{extra_fields}{Optionally extract additional columns from the 
object. Default \code{NULL}.}
}
\value{
Returns a data.table of variants from a maf which can be used to
create a \code{bagel} object.
}
\description{
Add Description
}
\examples{
maf_file <- system.file("testdata", "public_TCGA.LUSC.maf", package = "BAGEL")
library(maftools)
maf <- read.maf(maf_file)
variants <- extract_variants_from_maf(maf = maf)
}
