% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{find_signatures}
\alias{find_signatures}
\title{Discovers signatures and weights from a table of counts using NMF}
\usage{
find_signatures(
  input,
  table_name,
  num_signatures,
  method = "lda",
  seed = NULL,
  nstart = 1,
  par = FALSE
)
}
\arguments{
\item{input}{A bagel object or counts table}

\item{table_name}{Name of table used for signature discovery}

\item{num_signatures}{Number of signatures to discover, k}

\item{method}{Discovery of new signatures using either LDA or NMF}

\item{seed}{Seed for reproducible signature discovery}

\item{nstart}{Number of independent runs with optimal chosen (lda only)}

\item{par}{Number of parallel cores to use (NMF only)}
}
\value{
Returns a result object with results and input object (if bagel)
}
\description{
Discovers signatures and weights from a table of counts using NMF
}
\examples{
#print("test")
#a <- readRDS(system.file("testdata", "TODO_BAGEL.rds", package = "BAGEL"))
#b <- find_signatures(input = a, table_name = "SNV96", num_signatures = 3,
#method = "nmf", seed = 12345, nstart = 1)
}
