% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number-analysis.R
\name{fit_background_distribution}
\alias{fit_background_distribution}
\title{Fit Background Copy Number Distribution}
\usage{
fit_background_distribution(segments)
}
\arguments{
\item{segments}{Copy number segments (preferably from normal samples)}
}
\value{
Fitted background distribution parameters
}
\description{
Fit Background Copy Number Distribution
}
\keyword{internal}
