% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel-workflow.R
\name{generate_analysis_report}
\alias{generate_analysis_report}
\title{Generate Analysis Report}
\usage{
generate_analysis_report(cancer_type, results, output_dir)
}
\arguments{
\item{cancer_type}{Character, cancer type name}

\item{results}{List, analysis results}

\item{output_dir}{Character, output directory}
}
\description{
Generate Analysis Report
}
