% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{generate_result_grid}
\alias{generate_result_grid}
\title{Generate result_grid from bagel based on annotation and range of k}
\usage{
generate_result_grid(
  bagel,
  discovery_type = "lda",
  annotation,
  k_start,
  k_end,
  n_start = 1,
  seed = NA,
  verbose = FALSE
)
}
\arguments{
\item{bagel}{Input bagel to generate grid from}

\item{discovery_type}{Algorithm for signature discovery}

\item{annotation}{Sample annotation to split results into}

\item{k_start}{Lower range of number of signatures for discovery}

\item{k_end}{Upper range of number of signatures for discovery}

\item{seed}{Give a seed to generate reproducible signatures}

\item{verbose}{Whether to output loop iterations}

\item{num_iter}{Number of times to discover signatures and compare based on
loglikihood}
}
\value{
Results a result object containing signatures and sample weights
}
\description{
Generate result_grid from bagel based on annotation and range of k
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
generate_result_grid(bay, k_start = 2, k_end = 5)
}
