% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAGEL.R
\name{generate_simulate_data}
\alias{generate_simulate_data}
\title{The function generates simulated data}
\usage{
generate_simulate_data(
  D,
  Q,
  S,
  N,
  H,
  B = 10,
  rho = 0.1,
  prob = NULL,
  f = NULL,
  Comega = NULL,
  a = c(-100, 0, 1, 2, 100),
  sigma_square = 1,
  seed = 1
)
}
\arguments{
\item{D}{the number of drugs}

\item{Q}{the number of depression scores}

\item{S}{the number of covariates}

\item{N}{the number of patients in the dataset}

\item{H}{the number of clusters}

\item{B}{the number of bases for P-spline}

\item{rho}{the prior probability of the binary indicator cube entries being 1}

\item{prob}{the probabilities of patients being assigned to clusters. If not given, it will be equal probabilities.}

\item{f}{the function designed to capture the longitudinally on depression. If not given, it will be constant 0}

\item{Comega}{the correlation matrix which captures the dependencies among depression items. If not given, it will be an identity matrix}

\item{a}{the thresholds that connects the latent and observed depression scores}

\item{sigma_square}{the variance of noise for the latent continuous depression scores}

\item{seed}{the starting number used to generate random numbers}
}
\value{
a simulated dataset with observated data and unobserved parameters
}
\description{
This function generates simulated data, provided the scale of the dataset
}
\examples{
D = 5; Q=3; S=5; N=100;H=10
simulated_data <- generate_simulate_data(D, Q, S, N, H)
}
