% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint-similarity-analysis.R
\name{generate_statistical_report}
\alias{generate_statistical_report}
\title{Internal function to generate statistical report}
\usage{
generate_statistical_report(matched_data, stat_result, dataset_names)
}
\arguments{
\item{matched_data}{Matched breakpoint data}

\item{stat_result}{Statistical results}

\item{dataset_names}{Dataset names}
}
\value{
Character vector with report content
}
\description{
Internal function to generate statistical report
}
\keyword{internal}
