% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel-workflow.R
\name{get_arm_definitions}
\alias{get_arm_definitions}
\title{Get Arm Definitions for Cancer Type}
\usage{
get_arm_definitions(cancer_type, breakpoint_data)
}
\arguments{
\item{cancer_type}{Character, cancer type name}

\item{breakpoint_data}{List, loaded breakpoint data}
}
\value{
Arm definitions data frame
}
\description{
Get Arm Definitions for Cancer Type
}
