% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel-workflow.R
\name{get_available_cancer_types}
\alias{get_available_cancer_types}
\title{Get Available Cancer Types}
\usage{
get_available_cancer_types(data_dir = NULL)
}
\arguments{
\item{data_dir}{Character, path to data directory}
}
\value{
Character vector of available cancer types
}
\description{
Get Available Cancer Types
}
