% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_breakpoint_data}
\alias{get_breakpoint_data}
\title{Get Built-in Breakpoint Data}
\usage{
get_breakpoint_data(cancer_type = "consensus", data_type = "arm_definitions")
}
\arguments{
\item{cancer_type}{Character, specific cancer type or "consensus" for pan-cancer}

\item{data_type}{Character, type of breakpoint data to retrieve}
}
\value{
Breakpoint data frame
}
\description{
Get Built-in Breakpoint Data
}
