% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{get_sample_names}
\alias{get_sample_names}
\title{Return samples names for bagel object}
\usage{
get_sample_names(bay)
}
\arguments{
\item{bay}{Bagel object containing samples}
}
\value{
Returns names of samples in bagel object
}
\description{
Return samples names for bagel object
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
sample_names(bay)
}
