% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{get_variants}
\alias{get_variants}
\title{Return variants for bagel object}
\usage{
get_variants(bay)
}
\arguments{
\item{bay}{Bagel object containing variants}
}
\value{
Returnsvariants in bagel object
}
\description{
Return variants for bagel object
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
get_variants(bay)
}
