% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{head_and_neck}
\alias{head_and_neck}
\title{head_and_neck}
\format{
\subsection{head_and_neck}{

\describe{
\item{chr}{chromosome}
\item{arm}{chromosome arm}
\item{cytoband}{cytobands}
\item{start}{start base of breakpoint region}
\item{end}{end base of breakpoint region}
\item{direction}{direction of the breakpoint conferring advantage}
\item{telcent}{direction of the breakpoint calculated, telomere to centromere or opposite}
\item{id}{combined id of each breakpoint event}
}
}
}
\source{
\url{https://www.github.com/PollyHung/BAGEL/docs/head_and_neck.txt}
}
\usage{
head_and_neck
}
\description{
Breakpoint File for TCGA Head and Neck Cancer
}
\keyword{datasets}
