% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number-analysis.R
\name{identify_alterations}
\alias{identify_alterations}
\title{Identify Copy Number Alterations}
\usage{
identify_alterations(segments_annotated, amp_threshold, del_threshold)
}
\arguments{
\item{segments_annotated}{Segments with arm annotations}

\item{amp_threshold}{Amplification threshold}

\item{del_threshold}{Deletion threshold}
}
\value{
Segments with alteration classifications
}
\description{
Identify Copy Number Alterations
}
\keyword{internal}
