% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{init_sample_annotations}
\alias{init_sample_annotations}
\title{Initialize sample annotation data.table with sample names from variants}
\usage{
init_sample_annotations(bay)
}
\arguments{
\item{bay}{Bagel object we input sample into}
}
\value{
Sets sample_annotations slot {no return}
}
\description{
Initialize sample annotation data.table with sample names from variants
}
\examples{
bay <- readRDS(system.file("testdata", "bagel_sbs96.rds", package = "BAGEL"))
init_sample_annotations(bay)
bay
}
