% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSegs-fixed.R
\name{join_segments_improved}
\alias{join_segments_improved}
\title{Improved Segment Joining Logic}
\usage{
join_segments_improved(segments, coord, arm_type, telcent)
}
\arguments{
\item{segments}{Segments to join}

\item{coord}{Arm coordinates}

\item{arm_type}{p or q arm}

\item{telcent}{tel or cent}
}
\value{
List with length, start, end
}
\description{
Improved Segment Joining Logic
}
\keyword{internal}
