% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{lda_posterior}
\alias{lda_posterior}
\title{LDA prediction of samples based on existing signatures (default COSMIC)}
\usage{
lda_posterior(bagel, signatures = cosmic_result@signatures,
  signatures_to_use = seq_len(ncol(signatures)))
}
\arguments{
\item{bagel}{Input samples to predit signature weights}

\item{signatures}{Signatures to use for prediction (default COSMIC)}

\item{signatures_to_use}{Which signatures in set to use (default all)}
}
\value{
Results a result object containing signatures and sample weights
}
\description{
LDA prediction of samples based on existing signatures (default COSMIC)
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
lda_posterior(bay)
}
