% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel-workflow.R
\name{load_breakpoint_data}
\alias{load_breakpoint_data}
\title{Load Pan-Cancer Breakpoint Data}
\usage{
load_breakpoint_data()
}
\value{
List containing all breakpoint datasets
}
\description{
Load Pan-Cancer Breakpoint Data
}
