% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint-similarity-analysis.R
\name{load_query_data}
\alias{load_query_data}
\title{Internal function to load query data}
\usage{
load_query_data(query_path, dataset_name)
}
\arguments{
\item{query_path}{Path to query dataset file}

\item{dataset_name}{Name for the query dataset}
}
\value{
Data frame with query breakpoint data
}
\description{
Internal function to load query data
}
\keyword{internal}
