% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagel-workflow.R
\name{load_segmentation_data}
\alias{load_segmentation_data}
\title{Load and Clean Segmentation Data for a Cancer Type}
\usage{
load_segmentation_data(cancer_type, data_dir)
}
\arguments{
\item{cancer_type}{Character, cancer type name}

\item{data_dir}{Character, path to data directory}
}
\value{
Cleaned segmentation data frame
}
\description{
Load and Clean Segmentation Data for a Cancer Type
}
