% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{maf_file_to_dt}
\alias{maf_file_to_dt}
\title{Loads a maf file and extracts the data.table of variants}
\usage{
maf_file_to_dt(
  maf_file,
  filter = TRUE,
  only_snp = TRUE,
  used_fields = c("Chromosome", "Start_Position", "End_Position", "Tumor_Seq_Allele1",
    "Tumor_Seq_Allele2", "Tumor_Sample_Barcode")
)
}
\arguments{
\item{maf_file}{Location of maf file}

\item{filter}{Filter to only passed variants}

\item{only_snp}{Filter only non-snp variants}

\item{used_fields}{Which fields to extract}
}
\value{
Returns a data.table of variants from a maf
}
\description{
Loads a maf file and extracts the data.table of variants
}
\examples{
maf_file=system.file("testdata", "public_TCGA.LUSC.maf", package = "BAGEL")
maf = BAGEL::maf_file_to_dt(maf_file = maf_file)
}
