% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{maf_files_to_dt}
\alias{maf_files_to_dt}
\title{Loads a list of mafs and converts to a combined data.table}
\usage{
maf_files_to_dt(
  maf_files,
  filter = TRUE,
  only_snp = FALSE,
  extra_fields = NULL,
  verbose = FALSE
)
}
\arguments{
\item{maf_files}{A list of vcf file locations}

\item{filter}{Filter to only passed variants}

\item{only_snp}{Filter only non-snp variants}

\item{extra_fields}{Which additional fields to extract}

\item{verbose}{Show file list progress}
}
\value{
Returns a data.table of variants from mafs
}
\description{
Loads a list of mafs and converts to a combined data.table
}
\examples{
mafs <- list.files(system.file("testdata", package = "BAGEL"),
  pattern = glob2rx("*maf"), full.names = TRUE)
public_mafs <- BAGEL::maf_files_to_dt(maf_files = mafs)
}
