% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{maf_to_dt}
\alias{maf_to_dt}
\title{Extracts the data.table of variants from a maftools maf object}
\usage{
maf_to_dt(
  maf,
  maf_name = NULL,
  filter = TRUE,
  only_snp = TRUE,
  extra_fields = NULL
)
}
\arguments{
\item{maf}{Maf object loaded by maftools::read.maf()}

\item{maf_name}{Name of the sample or maf}

\item{filter}{Filter to only passed variants}

\item{only_snp}{Filter only non-snp variants}

\item{extra_fields}{Which additional fields to extract}
}
\value{
Returns a data.table of variants from a maf
}
\description{
Extracts the data.table of variants from a maftools maf object
}
\examples{
maf_file=system.file("testdata", "public_TCGA.LUSC.maf", package = "BAGEL")
maf = maftools::read.maf(maf_file)
maf_dt = BAGEL::maf_to_dt(maf = maf)
}
