% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMarkers.R
\name{makeMarkers}
\alias{makeMarkers}
\title{Make Markers from Bed file or from Gencode v47}
\usage{
makeMarkers(
  bed_file = NULL,
  technique = "TS",
  genome_build = "hg38",
  gff.path = NULL,
  marker.path = NULL
)
}
\arguments{
\item{bed_file}{A string representing the path to a BED file (optional). Used if the targeted
sequencing panel is not predefined.}

\item{technique}{A character string specifying the sequencing technique to use. Options are
"TS", "WGS", "WES", or "SNP".}

\item{genome_build}{A character string specifying the genome build to use (default is "hg38").}

\item{gff.path}{A string representing the path to a GFF file (optional). Used for WGS and WES
techniques if a custom GFF is required.}

\item{marker.path}{A string representing the path to a marker file (optional). Required for
SNP array technique.}

\item{panel}{A character string specifying the targeted gene panel to use (default is
"oncoKB_annot"). It should be one of the predefined panels.}
}
\value{
A GRanges object containing the regions of interest derived from the specified
sequencing technique.
}
\description{
This function estimates arm-level copy number based on various sequencing techniques,
including targeted sequencing (TS), whole genome sequencing (WGS), whole exome sequencing (WES),
and SNP arrays. It generates a GRanges object containing the regions of interest based on the
specified parameters.
}
\examples{
markers <- makeMarkers(technique = "WGS", gff.path = "path/to/gff")
}
