% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSegments.R
\name{mergeSegments}
\alias{mergeSegments}
\title{Merge Segments Function}
\usage{
mergeSegments(seg.path, min_probes = 10)
}
\arguments{
\item{seg.path}{A string representing the file path to a tab-delimited text file
containing the segmentation data.}

\item{min_probes}{An integer specifying the minimum number of probes required
to consider a segment as valid. Segments with fewer probes
than this threshold will be merged into the preceding segment.}
}
\value{
A data frame containing the merged segments, which includes the columns:
'End', 'Num_Probes', and 'Segment_Mean'.
}
\description{
This function merges segments based on the number of probes. Segments with fewer than a specified
number of probes are merged into the previous segment.
}
\details{
The function reads the segmentation data from the specified file path, merges segments based on the probe count,
and returns a data frame containing the merged segments with columns 'End', 'Num_Probes', and 'Segment_Mean'.
}
