% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint-similarity-analysis.R
\name{perform_similarity_analysis_internal}
\alias{perform_similarity_analysis_internal}
\title{Internal function to perform similarity analysis on combined data}
\usage{
perform_similarity_analysis_internal(combined_data, dataset_names)
}
\arguments{
\item{combined_data}{Combined breakpoint data}

\item{dataset_names}{Names of the datasets}
}
\value{
Data frame with matched breakpoints
}
\description{
Internal function to perform similarity analysis on combined data
}
\keyword{internal}
