% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint-similarity-analysis.R
\name{perform_statistical_tests_internal}
\alias{perform_statistical_tests_internal}
\title{Internal function to perform statistical tests}
\usage{
perform_statistical_tests_internal(matched_data)
}
\arguments{
\item{matched_data}{Matched breakpoint data}
}
\value{
List with statistical results
}
\description{
Internal function to perform statistical tests
}
\keyword{internal}
