% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_exposures}
\alias{plot_exposures}
\title{Plots signature weights for each sample}
\usage{
plot_exposures(
  result,
  proportional = TRUE,
  label_samples = TRUE,
  samples_plotted = colnames(result@samples),
  sort_samples = "numerical",
  num_samples = length(colnames(result@samples))
)
}
\arguments{
\item{result}{S4 Result Object}

\item{proportional}{Whether weights are normalized to sum to 1 or not}

\item{label_samples}{Whether to display sample names or not}

\item{samples_plotted}{A subset of samples to plot}

\item{sort_samples}{Defaults to numerical, may be total 'counts', or a
specific signatures name (e.g. 'Signature1)}

\item{num_samples}{Number of sorted samples to plot}
}
\value{
Generates plot {no return}
}
\description{
Plots signature weights for each sample
}
\examples{
result <- readRDS(system.file("testdata", "res.rds", package = "BAGEL"))
plot_exposures(result)
}
