% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_exposures_by_annotation}
\alias{plot_exposures_by_annotation}
\title{Plots signature weights for each sample gridded by single annotation}
\usage{
plot_exposures_by_annotation(
  result,
  proportional = TRUE,
  label_samples = FALSE,
  sort_samples = "numerical",
  annotation,
  by_group = TRUE
)
}
\arguments{
\item{result}{S4 Result Object}

\item{proportional}{Whether weights are normalized to sum to 1 or not}

\item{label_samples}{Whether to display sample names or not}

\item{sort_samples}{Defaults to numerical, may be total 'counts', or a
specific signatures name (e.g. 'Signature1)}

\item{annotation}{Annotation to use for facet_wrap}

\item{by_group}{Plot subplots by annotation or by signature}
}
\value{
Generates plot {no return}
}
\description{
Plots signature weights for each sample gridded by single annotation
}
\examples{
result <- readRDS(system.file("testdata", "res.rds", package = "BAGEL"))
plot_exposures(result)
}
