% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sample}
\alias{plot_sample}
\title{Return sample from bagel object}
\usage{
plot_sample(bay, sample_name)
}
\arguments{
\item{bay}{Bagel object containing samples}

\item{sample_name}{Sample name to plot counts}
}
\value{
Generates sample plot {no return}
}
\description{
Return sample from bagel object
}
