% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sample_counts}
\alias{plot_sample_counts}
\title{Return sample from bagel object}
\usage{
plot_sample_counts(bay, table_name, sample_name)
}
\arguments{
\item{bay}{Bagel object containing samples}

\item{table_name}{Name of table used for plotting counts}

\item{sample_name}{Sample name to plot counts}
}
\value{
Generates sample plot {no return}
}
\description{
Return sample from bagel object
}
\examples{
bay <- readRDS(system.file("testdata", "bagel_sbs96.rds", package = "BAGEL"))
plot_sample_counts(bay, "SBS96", get_sample_names(bay)[1])
}
