% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sample_reconstruction_error}
\alias{plot_sample_reconstruction_error}
\title{Plotting Signature Motif counts/spectra}
\usage{
plot_sample_reconstruction_error(result, sample_number)
}
\arguments{
\item{result}{S4 Result Object}
}
\value{
Generates plot {no return}
}
\description{
Plotting Signature Motif counts/spectra
}
\examples{
result <- readRDS(system.file("testdata", "res.rds", package = "BAGEL"))
plot_signatures(result)
}
