% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_samples}
\alias{plot_samples}
\title{Plots signature weights for each sample}
\usage{
plot_samples(result, proportional = TRUE)
}
\arguments{
\item{result}{S4 Result Object}

\item{proportional}{Whether weights are normalized to sum to 1 or not}
}
\value{
Generates plot {no return}
}
\description{
Plots signature weights for each sample
}
\examples{
result <- readRDS(system.file("testdata", "res.rds", package = "BAGEL"))
plot_samples(result)
}
