% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCopyNumber-fixed.R
\name{process_breakpoint_input}
\alias{process_breakpoint_input}
\title{Process Breakpoint Input}
\usage{
process_breakpoint_input(breakpoints)
}
\arguments{
\item{breakpoints}{Input breakpoints in various formats}
}
\value{
Standardized arm definitions
}
\description{
Handle different formats of breakpoint input and return standardized format
}
\keyword{internal}
