% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCopyNumber-fixed.R
\name{save_analysis_results}
\alias{save_analysis_results}
\title{Save Analysis Results to Files}
\usage{
save_analysis_results(results, output_dir, cancer_type)
}
\arguments{
\item{results}{Analysis results}

\item{output_dir}{Output directory}

\item{cancer_type}{Cancer type}
}
\description{
Save Analysis Results to Files
}
\keyword{internal}
