% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{select_genome}
\alias{select_genome}
\title{Helper function to load HG19 or HG38}
\usage{
select_genome(hg)
}
\arguments{
\item{hg}{Which human genome build number to return}
}
\value{
Returns BSgenome of given version
}
\description{
Helper function to load HG19 or HG38
}
\examples{
g <- select_genome(hg = 38)

}
