% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{set_sample_annotations}
\alias{set_sample_annotations}
\title{Set sample level annotations for bagel object}
\usage{
set_sample_annotations(bay, annotations)
}
\arguments{
\item{bay}{Bagel object we input sample into}

\item{annotations}{Sample DataFrame}
}
\value{
Sets sample_annotations slot {no return}
}
\description{
Set sample level annotations for bagel object
}
\examples{
annotations <- readRDS(system.file("testdata", "dt.rds", package = "BAGEL"))
bay <- new("bagel")
set_sample_annotations(bay, annotations)
}
