% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{set_samples}
\alias{set_samples}
\title{Set samples for bagel object}
\usage{
set_samples(bay, samp)
}
\arguments{
\item{bay}{Bagel object we input sample into}

\item{samp}{Sample DataFrame}
}
\value{
Sets sample slot {no return}
}
\description{
Set samples for bagel object
}
\examples{
samp <- readRDS(system.file("testdata", "dt.rds", package = "BAGEL"))
bay <- new("bagel")
set_samples(bay, dat)
}
