% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{set_variants}
\alias{set_variants}
\title{Set variants table for bagel object}
\usage{
set_variants(bay, variants)
}
\arguments{
\item{bay}{Bagel object we input sample into}

\item{variants}{Variant DataFrame}
}
\value{
Sets variant slot {no return}
}
\description{
Set variants table for bagel object
}
\examples{
variants <- readRDS(system.file("testdata", "dt.rds", package = "BAGEL"))
bay <- new("bagel")
set_variants(bay, variants)
}
