% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSegs-fixed.R
\name{should_flip_coordinates}
\alias{should_flip_coordinates}
\title{Determine if coordinates should be flipped}
\usage{
should_flip_coordinates(arm_type, telcent)
}
\arguments{
\item{arm_type}{p or q}

\item{telcent}{tel or cent}
}
\value{
Logical
}
\description{
Determine if coordinates should be flipped
}
\keyword{internal}
