% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSegs-fixed.R
\name{should_trim_from_end}
\alias{should_trim_from_end}
\title{Determine if trimming should happen from end}
\usage{
should_trim_from_end(arm_type, telcent)
}
\arguments{
\item{arm_type}{p or q}

\item{telcent}{tel or cent}
}
\value{
Logical
}
\description{
Determine if trimming should happen from end
}
\keyword{internal}
