\name{simulated_data}
\alias{simulated_data}
\docType{data}
\title{An example simulated dataset.}
\description{
  It is a list generated by function \link{generate_simulate_data} (the example code). We assumed 5 drugs, 3 depression items, 5 covariates, 10 bases, and 100 patients.
}
\usage{data(simulated_data)}
\format{
  A list giving the following variables.
  \describe{
    \item{\code{U}}{the depression scores, each row corresponds to one visit}
    \item{\code{Z}}{the binary indicator matrix for drug usage information, each row corresponds to one visit}
    \item{\code{X}}{the covariates matrix, it also includes the patient ID (1st column) and visit time (2nd column). Each row corresponds to one visit}
    \item{\code{D}}{the number of drugs}
    \item{\code{Q}}{the number of depression scores}
    \item{\code{S}}{the number of covariates}
    \item{\code{N}}{the number of patients in the dataset}
    \item{\code{B}}{the number of bases for P-spline}
    \item{\code{H}}{the number of clusters}
    \item{\code{beta}}{the linear coefficients not dependent on drug effect}
    \item{\code{alpha}}{the linear coefficients dependent on drug effect}
    \item{\code{gamma}}{the linear coefficients for the P-spline}
    \item{\code{R}}{the binary indicator cube that introduce sparsity}
    \item{\code{e}}{the clustering membership vector}
    \item{\code{a}}{the thresholds that connects the latent and observed depression scores}
    \item{\code{Comega}}{the correlation matrix which captures the dependencies among depression items}
  }
}

\examples{
  # output true clustering membership
  data("simulated_data")
  simulated_data$e

  # get the details of the list
  str(simulated_data)
}
\keyword{datasets}
