% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number-analysis.R
\name{split_multi_arm_segments}
\alias{split_multi_arm_segments}
\title{Split Segments Spanning Multiple Arms}
\usage{
split_multi_arm_segments(segments_annotated, arm_definitions)
}
\arguments{
\item{segments_annotated}{Segments with arm annotations}

\item{arm_definitions}{Arm definitions}
}
\value{
Segments with proper arm splits
}
\description{
Split Segments Spanning Multiple Arms
}
\keyword{internal}
