% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{subset_bagel_by_annotation}
\alias{subset_bagel_by_annotation}
\title{Creates a new bagel subsetted to only one value of a sample annotation}
\usage{
subset_bagel_by_annotation(bay, annot_col, annot_names)
}
\arguments{
\item{bay}{Input bagel}

\item{annot_col}{Annotation class to use for subsetting}

\item{annot_names}{Annotational value to subset to}
}
\value{
Returns a new bagel object with sample annotations, count tables,
and variants subsetted to only contains samples of the specified annotation
type
}
\description{
Creates a new bagel subsetted to only one value of a sample annotation
}
\examples{
bay <- readRDS(system.file("testdata", "bagel_sbs96.rds", package = "BAGEL"))
sample_annotations <- read.table(system.file("testdata",
"sample_annotations.txt", package = "BAGEL"), sep = "\t", header=TRUE)
init_sample_annotations(bay)
add_sample_annotations(bay, sample_annotations, sample_column =
"Sample_Names", columns_to_add = "Tumor_Subtypes")
subset_bagel_by_annotation(bay, "Tumor_Subtypes", "Lung")
}
