% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{subset_bagel_by_counts}
\alias{subset_bagel_by_counts}
\title{Creates a new bagel subsetted to only samples with enough variants}
\usage{
subset_bagel_by_counts(bay, table_name, num_counts)
}
\arguments{
\item{bay}{Input bagel}

\item{table_name}{Name of table used for subsetting}

\item{num_counts}{Minimum sum count value to drop samples}
}
\description{
Creates a new bagel subsetted to only samples with enough variants
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
subset_bagel_by_counts(bay, "SNV96", 5)
}
