% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{subset_variants_by_samples}
\alias{subset_variants_by_samples}
\title{Return sample from bagel object}
\usage{
subset_variants_by_samples(bay, sample_name)
}
\arguments{
\item{bay}{Bagel object containing samples}

\item{sample_name}{Sample name to subset by}
}
\value{
Returns sample dataframe subset to a single sample
}
\description{
Return sample from bagel object
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
subset_variants_by_samples(bay, "public_LUAD_TCGA-97-7938.vcf")
}
