% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_breakpoints}
\alias{validate_breakpoints}
\title{Validate Breakpoint Data Structure}
\usage{
validate_breakpoints(breakpoints)
}
\arguments{
\item{breakpoints}{Breakpoint data structure}
}
\value{
Validated breakpoints or throws error
}
\description{
Validate Breakpoint Data Structure
}
\keyword{internal}
