% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_gistic2_parameters}
\alias{validate_gistic2_parameters}
\title{Validate GISTIC2 Parameters}
\usage{
validate_gistic2_parameters(amp_threshold, del_threshold, stringent_threshold)
}
\arguments{
\item{amp_threshold}{Amplification threshold}

\item{del_threshold}{Deletion threshold}

\item{stringent_threshold}{Stringent threshold}
}
\value{
Validated parameters or throws error
}
\description{
Validate GISTIC2 Parameters
}
