% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_segments}
\alias{validate_segments}
\title{Validate Segmentation Data Frame}
\usage{
validate_segments(segments)
}
\arguments{
\item{segments}{Data frame with segmentation data}
}
\value{
Validated data frame or throws error
}
\description{
Validate Segmentation Data Frame
}
\keyword{internal}
