% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{vcf_to_dt}
\alias{vcf_to_dt}
\title{Converts a loaded vcf object to data.table}
\usage{
vcf_to_dt(
  vcf,
  vcf_name = NULL,
  filter = TRUE,
  only_snp = FALSE,
  extra_fields = NULL
)
}
\arguments{
\item{vcf}{Location of vcf file}

\item{vcf_name}{Name of the sample or vcf}

\item{filter}{Filter to only passed variants}

\item{only_snp}{Filter only non-snp variants}

\item{extra_fields}{Which additional fields to extract}
}
\value{
Returns a data.table of variants from a vcf
}
\description{
Converts a loaded vcf object to data.table
}
\examples{
luad_vcf_file <- system.file("testdata", "public_LUAD_TCGA-97-7938.vcf",
  package = "BAGEL")
luad_vcf <- VariantAnnotation::readVcf(luad_vcf_file)
luad_vcf_name <- basename(luad_vcf_file)
luad <- BAGEL::vcf_to_dt(vcf = luad_vcf, vcf_name = luad_vcf_name)
}
