% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{vcfs_to_dt}
\alias{vcfs_to_dt}
\title{Loads a list of vcfs and converts to a combined data.table}
\usage{
vcfs_to_dt(vcfs)
}
\arguments{
\item{vcfs}{A list of vcf file locations}
}
\value{
Returns a data.table of variants from vcfs
}
\description{
Loads a list of vcfs and converts to a combined data.table
}
\examples{
melanoma_vcf <- list.files(system.file("testdata", package = "MotifSig"),
  pattern = glob2rx("*SKCM*vcf"), full.names = TRUE)
melanoma <- MotifSig::vcfs_to_dt(vcfs = melanoma_vcf)
}
