% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WES}
\alias{WES}
\title{WES}
\format{
\subsection{WES}{

\describe{
\item{breakpoint_ids}{breakpoint_ids}
}
}
}
\source{
\url{https://www.gencodegenes.org/human/}
}
\usage{
WES
}
\description{
Gencode v47 Whole Exomes
}
\keyword{datasets}
