% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WGS}
\alias{WGS}
\title{WGS}
\format{
\subsection{WGS}{

\describe{
\item{seqnames}{chromosomes}
\item{ranges}{start-end}
\item{strand}{* or +/-}
}
}
}
\source{
\url{https://www.gencodegenes.org/human/}
}
\usage{
WGS
}
\description{
GRange object, Gencode v47 Whole Genomes
}
\keyword{datasets}
