% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{what_cosmic30_sigs}
\alias{what_cosmic30_sigs}
\title{Input a cancer subtype to return a list of related COSMIC signatures}
\usage{
what_cosmic30_sigs(tumor_type)
}
\arguments{
\item{tumor_type}{Cancer subtype to view related signatures}
}
\value{
Returns signatures related to a partial string match
}
\description{
Input a cancer subtype to return a list of related COSMIC signatures
}
\examples{
what_cosmic30_sigs("lung")
}
