% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{add_flank_to_variants}
\alias{add_flank_to_variants}
\title{Uses a genome object to find context and add it to the variant table}
\usage{
add_flank_to_variants(bay, g, flank_start, flank_end)
}
\arguments{
\item{bay}{Input samples}

\item{g}{Genome object used for finding variant context}

\item{flank_start}{Start of flank area to add, can be positive or negative}

\item{flank_end}{End of flank area to add, can be positive or negative}
}
\description{
Uses a genome object to find context and add it to the variant table
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
g <- select_genome("38")
add_flank_to_variants(bay, g, 1, 5)
add_flank_to_variants(bay, g, -5, -1)
}
