% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{add_sample_annotations}
\alias{add_sample_annotations}
\title{Adds sample annotation to bagel object with available samples}
\usage{
add_sample_annotations(
  bay,
  annotations,
  sample_column = Sample_ID,
  columns_to_add
)
}
\arguments{
\item{bay}{Bagel object we input sample into}
}
\value{
Sets sample_annotations slot {no return}
}
\description{
Adds sample annotation to bagel object with available samples
}
\examples{
annotations <- readRDS(system.file("testdata", "dt.rds", package = "BAGEL"))
bay <- new("bagel")
init_sample_annotations(bay)
}
