% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{Create Background Files from Breakpoint Data}
\usage{
background(rd = result_dir)
}
\arguments{
\item{result_dir}{A string specifying the directory where the breakpoint
files are located and where the output background files will be saved.}
}
\value{
NULL; the function saves two files: one for telomere and one for
centromere backgrounds in the specified result directory. The function
outputs messages indicating the progress of file creation.
}
\description{
This function processes breakpoint files generated during preprocessing,
categorizing them into telomere and centromere segments. It also creates
a directory to store the output files if it does not already exist.
}
\examples{
# Example usage
background(result_dir = "path/to/directory")

}
