% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\docType{class}
\name{bagel-class}
\alias{bagel-class}
\title{The primary object for BAGEL that contains all samples and tables}
\description{
The primary object for BAGEL that contains all samples and tables
}
\section{Slots}{

\describe{
\item{\code{variants}}{A list of variants objects containing variant-level information}

\item{\code{counts_table}}{Summary tables with unnormalized motif counts}
}}

