% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{build_standard_table}
\alias{build_standard_table}
\title{Builds a standard table from user variants}
\usage{
build_standard_table(bay, g, table_name, strand_type = NA, overwrite = FALSE)
}
\arguments{
\item{bay}{Input samples}

\item{g}{A \linkS4class{BSgenome} object indicating which genome
reference the variants and their coordinates were derived from.}

\item{table_name}{Name of standard table to build SBS96, SBS192, DBS, or
Indel}

\item{strand_type}{Only for SBS192 Transcript_Strand or Replication_Strand}

\item{overwrite}{Overwrite existing count table}
}
\value{
None
}
\description{
Builds a standard table from user variants
}
\examples{
g <- select_genome("19")

bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
build_standard_table(bay, g, "SBS96", overwrite = TRUE)

bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
annotate_transcript_strand(bay, "19")
build_standard_table(bay, g, "SBS192", "Transcript_Strand")

bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
annotate_replication_strand(bay, BAGEL::rep_range)
build_standard_table(bay, g, "SBS192", "Replication_Strand")

bay <- readRDS(system.file("testdata", "dbs_bagel.rds",
package = "BAGEL"))
build_standard_table(bay, g, "DBS")

bay <- readRDS(system.file("testdata", "indel_bagel.rds", package = "BAGEL"))
build_standard_table(bay, g, table_name = "INDEL")
}
