% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{compare_results}
\alias{compare_results}
\title{Compare two result files or input one to compare to COSMIC}
\usage{
compare_results(
  result,
  other_result = cosmic_v2_sigs,
  threshold = 0.9,
  result_name = "User Signatures 1",
  other_result_name = "User Signatures 2"
)
}
\arguments{
\item{result}{Result to compare}

\item{other_result}{Second result, leave blank to use cosmic results}

\item{threshold}{threshold for similarity}

\item{result_name}{title for plot of first result signatures}

\item{other_result_name}{title for plot of second result signatures}
}
\value{
Returns the comparisons
}
\description{
Compare two result files or input one to compare to COSMIC
}
\examples{
res <- readRDS(system.file("testdata", "res.rds", package = "BAGEL"))
compare_results(res, threshold = 0.8)
}
