% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_sbs96_table}
\alias{create_sbs96_table}
\title{Uses a genome object to find context and generate standard SBS96 tables}
\usage{
create_sbs96_table(bay, overwrite = FALSE)
}
\arguments{
\item{bay}{Input samples}

\item{overwrite}{Overwrite existing count table}
}
\description{
Uses a genome object to find context and generate standard SBS96 tables
}
