% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_snv192_table}
\alias{create_snv192_table}
\title{Uses a genome object to find context and generate standard SNV192 table
using transcript strand}
\usage{
create_snv192_table(bay, g)
}
\arguments{
\item{bay}{Input samples}

\item{g}{Genome object used for finding variant context}
}
\description{
Uses a genome object to find context and generate standard SNV192 table
using transcript strand
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
g <- select_genome("38")
annotate_transcript_strand(bay, "19")
create_snv192_table(bay, g)
}
