% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_snv192_table}
\alias{create_snv192_table}
\title{Uses a genome object to find context and generate standard SNV192 table
using transcript strand}
\usage{
create_snv192_table(bay, g, strand_type)
}
\arguments{
\item{bay}{Input samples}

\item{g}{Genome object used for finding variant context}

\item{strand_type}{Transcript_Strand or Replication_Strand}
}
\description{
Uses a genome object to find context and generate standard SNV192 table
using transcript strand
}
