% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_snv96_table}
\alias{create_snv96_table}
\title{Uses a genome object to find context and generate standard SNV96 tables}
\usage{
create_snv96_table(bay)
}
\arguments{
\item{bay}{Input samples}

\item{g}{Genome object used for finding variant context}
}
\description{
Uses a genome object to find context and generate standard SNV96 tables
}
