% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{dt_to_bagel_dt}
\alias{dt_to_bagel_dt}
\title{Converts a data.table to a filtered BAGEL data.table}
\usage{
dt_to_bagel_dt(
  dt,
  dt_name = NULL,
  file_type = "data.table",
  filter = TRUE,
  only_snp = TRUE,
  used_fields = c("Chromosome", "Start_Position", "End_Position", "Tumor_Seq_Allele1",
    "Tumor_Seq_Allele2", "Tumor_Sample_Barcode")
)
}
\arguments{
\item{dt}{data.table input by user or vcf or maf functions to be filtered}

\item{dt_name}{Name of the vcf or maf}

\item{file_type}{Type of file that was input}

\item{filter}{Filter to only passed variants}

\item{only_snp}{Filter only non-snp variants}

\item{used_fields}{Which fields to extract}
}
\value{
Returns a data.table of variants from a maf
}
\description{
Converts a data.table to a filtered BAGEL data.table
}
\examples{
maf_file=system.file("testdata", "public_TCGA.LUSC.maf", package = "BAGEL")
maf = maftools::read.maf(maf_file)
dt = BAGEL::maf_to_dt(maf)
maf_dt = BAGEL::dt_to_bagel_dt(maf = maf)
}
