% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{find_signatures}
\alias{find_signatures}
\title{Discovers signatures and weights from a table of counts using NMF}
\usage{
find_signatures(input, num_signatures, method = "lda")
}
\arguments{
\item{input}{A bagel object or counts table}

\item{num_signatures}{Number of signatures to discover, k}
}
\value{
Returns a result object with results and input object (if bagel)
}
\description{
Discovers signatures and weights from a table of counts using NMF
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
find_signatures(bay, num_signatures = 4)
}
