% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{get_sample_annotations}
\alias{get_sample_annotations}
\title{Return sample annotation from bagel object}
\usage{
get_sample_annotations(bay)
}
\arguments{
\item{bay}{Bagel object we input sample into}
}
\value{
Sets sample_annotations slot {no return}
}
\description{
Return sample annotation from bagel object
}
\examples{
annotations <- readRDS(system.file("testdata", "dt.rds", package = "BAGEL"))
bay <- new("bagel")
init_sample_annotations(bay)
show_sample_annotations(bay)
}
