% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_signatures.R
\name{infer_signatures}
\alias{infer_signatures}
\title{LDA prediction of samples based on existing signatures (default COSMIC)}
\usage{
infer_signatures(
  bagel,
  signatures = cosmic_result@signatures,
  signatures_to_use = seq_len(ncol(signatures)),
  verbose = FALSE
)
}
\arguments{
\item{bagel}{Input samples to predit signature weights}

\item{signatures}{Signatures to use for prediction (default COSMIC)}

\item{signatures_to_use}{Which signatures in set to use (default all)}

\item{verbose}{Whether to show intermediate results}
}
\value{
Results a result object containing signatures and sample weights
}
\description{
LDA prediction of samples based on existing signatures (default COSMIC)
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
infer_signatures(bay)
}
