% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSegs.R
\name{joinSegs}
\alias{joinSegs}
\title{Join Segments Function}
\usage{
joinSegs(segdf, TELCENT)
}
\arguments{
\item{segdf}{A data frame containing segment information, including Sample, Start, End, and other numeric columns.}

\item{TELCENT}{A string indicating segments to be considered for merging (e.g., "TEL" or "CENT").}
}
\value{
A data frame with merged segment information per sample, including Percent, Start, and End.
}
\description{
This function processes a data frame of segments, merging them based on specified criteria.
It looks for TEL or CENT segments and merges them accordingly.
}
