% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_bagel.R
\name{log_bagel}
\alias{log_bagel}
\title{Setup Logging System}
\usage{
log_bagel(log_level = "INFO", log_file = NULL)
}
\arguments{
\item{log_level}{Character, logging level ("DEBUG", "INFO", "WARN", "ERROR")}

\item{log_file}{Character, optional file path for logging}
}
\description{
Setup Logging System
}
