% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{maf_to_dt}
\alias{maf_to_dt}
\title{Loads a maf file and extracts the data.table of variants}
\usage{
maf_to_dt(maf_file)
}
\arguments{
\item{maf_file}{Location of maf file}
}
\value{
Returns a data.table of variants from a maf
}
\description{
Loads a maf file and extracts the data.table of variants
}
\examples{
maf_file=system.file("testdata", "public_TCGA.LUSC.maf", package = "BAGEL")
maf = BAGEL::maf_to_dt(maf_file = maf_file)
}
