% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_signatures}
\alias{plot_signatures}
\title{Plotting Signature Motif counts/spectra}
\usage{
plot_signatures(result, no_legend = FALSE, plotly = FALSE)
}
\arguments{
\item{result}{S4 Result Object}

\item{no_legend}{Remove legend from plot}

\item{plotly}{add plotly layer for plot interaction}
}
\value{
Generates plot {no return}
}
\description{
Plotting Signature Motif counts/spectra
}
\examples{
result <- readRDS(system.file("testdata", "res.rds", package = "BAGEL"))
plot_signatures(result)
}
