% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_prediction.R
\name{predict_exposure}
\alias{predict_exposure}
\title{LDA prediction of samples based on existing signatures}
\usage{
predict_exposure(
  bagel,
  table_name,
  signature_res,
  signatures_to_use = seq_len(ncol(signature_res@signatures)),
  verbose = FALSE
)
}
\arguments{
\item{bagel}{Input samples to predit signature weights}

\item{table_name}{Name of table used for posterior prediction.
Must match the table type used to generate the prediction signatures}

\item{signature_res}{Signatures to use for prediction}

\item{signatures_to_use}{Which signatures in set to use (default all)}

\item{verbose}{Whether to show intermediate results}
}
\value{
Results a result object containing signatures and sample weights
}
\description{
LDA prediction of samples based on existing signatures
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
g <- select_genome("19")
build_standard_table(bay, g, "SNV96")
predict_exposure(bay, "SNV96", BAGEL::cosmic_v2_sigs)
}
