% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{subset_bagel_by_annotation}
\alias{subset_bagel_by_annotation}
\title{Creates a new bagel subsetted to only one value of a sample annotation}
\usage{
subset_bagel_by_annotation(bay, annot_col, annot_name)
}
\arguments{
\item{bay}{Input bagel}

\item{annot_col}{Annotation class to use for subsetting}

\item{annot_name}{Annotational value to subset to}
}
\description{
Creates a new bagel subsetted to only one value of a sample annotation
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
subset_bagel_by_annotation(bay, "Tumor_Type", "Lung")
}
