% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{subset_bagel_by_counts}
\alias{subset_bagel_by_counts}
\title{Creates a new bagel subsetted to only samples with enough variants}
\usage{
subset_bagel_by_counts(bay, table_name, num_counts)
}
\arguments{
\item{bay}{Input bagel}

\item{table_name}{Name of table used for subsetting}

\item{num_counts}{Minimum sum count value to drop samples}
}
\value{
Returns a new bagel object with sample annotations, count tables,
and variants subsetted to only contains samples with the specified minimum
number of counts (column sums) in the specified table
}
\description{
Creates a new bagel subsetted to only samples with enough variants
}
\examples{
bay <- readRDS(system.file("testdata", "bagel_sbs96.rds", package = "BAGEL"))
subset_bagel_by_counts(bay, "SBS96", 20)
}
