% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{subset_variant_by_type}
\alias{subset_variant_by_type}
\title{Subsets a variant table based on Variant Type}
\usage{
subset_variant_by_type(tab, type)
}
\arguments{
\item{tab}{Input variant table}

\item{type}{Variant type to return e.g. "SNV", "INS", "DEL", "DBS"}
}
\description{
Subsets a variant table based on Variant Type
}
\examples{
bay <- readRDS(system.file("testdata", "bagel.rds", package = "BAGEL"))
annotate_variant_type(bay)
subset_variant_by_type(get_variants(bay), "SNV")
}
