% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{vcf_file_to_dt}
\alias{vcf_file_to_dt}
\title{Loads a vcf and converts to data.table}
\usage{
vcf_file_to_dt(
  vcf_file,
  filter = TRUE,
  only_snp = TRUE,
  used_fields = c("Chromosome", "Start_Position", "End_Position", "Tumor_Seq_Allele1",
    "Tumor_Seq_Allele2", "Tumor_Sample_Barcode"),
  auto_fix_errors = TRUE
)
}
\arguments{
\item{vcf_file}{Location of vcf file}

\item{filter}{Filter to only passed variants}

\item{only_snp}{Filter only non-snp variants}

\item{used_fields}{Which fields to extract}
}
\value{
Returns a data.table of variants from a vcf
}
\description{
Loads a vcf and converts to data.table
}
\examples{
luad_vcf <- system.file("testdata", "public_LUAD_TCGA-97-7938.vcf",
  package = "BAGEL")
luad <- BAGEL::vcf_file_to_dt(vcf_file = luad_vcf)
}
