% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{vcf_files_to_dt}
\alias{vcf_files_to_dt}
\title{Loads a list of vcfs and converts to a combined data.table}
\usage{
vcf_files_to_dt(
  vcf_files,
  filter = TRUE,
  only_snp = TRUE,
  used_fields = c("Chromosome", "Start_Position", "End_Position", "Tumor_Seq_Allele1",
    "Tumor_Seq_Allele2", "Tumor_Sample_Barcode"),
  verbose = FALSE
)
}
\arguments{
\item{vcf_files}{A list of vcf file locations}

\item{filter}{Filter to only passed variants}

\item{only_snp}{Filter only non-snp variants}

\item{used_fields}{Which fields to extract}
}
\value{
Returns a data.table of variants from vcfs
}
\description{
Loads a list of vcfs and converts to a combined data.table
}
\examples{
melanoma_vcfs <- list.files(system.file("testdata", package = "BAGEL"),
  pattern = glob2rx("*SKCM*vcf"), full.names = TRUE)
melanoma <- BAGEL::vcf_files_to_dt(vcf_files = melanoma_vcfs)
}
