% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{vcf_to_dt}
\alias{vcf_to_dt}
\title{Loads a vcf and converts to data.table}
\usage{
vcf_to_dt(vcf_file)
}
\arguments{
\item{vcf_file}{Location of vcf file}
}
\value{
Returns a data.table of variants from a vcf
}
\description{
Loads a vcf and converts to data.table
}
\examples{
luad1_vcf <- system.file("testdata", "public_LUAD_TCGA-97-7938.vcf",
  package = "BAGEL")
luad1 <- BAGEL::vcf_to_dt(vcf = luad1_vcf)
}
