% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WES}
\alias{WES}
\title{WES}
\format{
\subsection{WES}{

\describe{
\item{seqnames}{chromosomes}
\item{ranges}{start-end}
\item{strand}{* or +/-}
}
}
}
\source{
\url{https://www.gencodegenes.org/human/}
}
\usage{
WES
}
\description{
GRange object, Gencode v47 Whole Exomes
}
\keyword{datasets}
