% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WGS}
\alias{WGS}
\title{WGS}
\format{
\subsection{WGS}{

\describe{
\item{breakpoint_ids}{breakpoint_ids}
}
}
}
\source{
\url{https://www.gencodegenes.org/human/}
}
\usage{
WGS
}
\description{
Gencode v47 Whole Genomes
}
\keyword{datasets}
